+++ tests/ui/ui3/java-testsuite-runner-ui3/CHANGELOG.rst
@@ -6,6 +6,18 @@ Changelog <https://keepachangelog.com/en/1.0.0/>`__, and this project
 adheres to `Semantic
 Versioning <https://semver.org/spec/v2.0.0.html>`__.
 
+
+.. _110--Unreleased:
+
+[1.1.0] - Unreleased
+--------------------
+
+Changed
+~~~~~~~
+
+- Rename the project to harmonize with other test suite projects (core, fs)
+- Make project "MicroEJ Platform Testsuite" compatible.
+
 .. _100---2021-04-20:
 
 [1.0.0] - 2021-04-20
@@ -18,6 +30,5 @@ Added
 -  Drawing time: retrieve the maximum drawing time to have the better framerate.
 
 ..
-    Copyright 2021 MicroEJ Corp. All rights reserved.
-    This library is provided in source code for use, modification and test, subject to license terms.
-    Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
+    Copyright 2021-2022 MicroEJ Corp. All rights reserved.
+    Use of this source code is governed by a BSD-style license that can be found with this software.